Adding CSS extraction to FormControl
