+++ packages/mui-material/src/FormControl/FormControl.js
@@ -3,13 +3,14 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import clsx from 'clsx';
 import composeClasses from '@mui/utils/composeClasses';
-import useThemeProps from '../styles/useThemeProps';
-import styled from '../styles/styled';
+import { styled, createUseThemeProps } from '../zero-styled';
 import { isFilled, isAdornedStart } from '../InputBase/utils';
 import capitalize from '../utils/capitalize';
 import isMuiElement from '../utils/isMuiElement';
 import FormControlContext from './FormControlContext';
 import { getFormControlUtilityClasses } from './formControlClasses';
+
+const useThemeProps = createUseThemeProps('MuiFormControl');
 
 const useUtilityClasses = (ownerState) => {
   const { classes, margin, fullWidth } = ownerState;
@@ -30,7 +31,7 @@ const FormControlRoot = styled('div', {
       ...(ownerState.fullWidth && styles.fullWidth),
     };
   },
-})(({ ownerState }) => ({
+})({
   display: 'inline-flex',
   flexDirection: 'column',
   position: 'relative',
@@ -40,18 +41,29 @@ const FormControlRoot = styled('div', {
   margin: 0,
   border: 0,
   verticalAlign: 'top', // Fix alignment issue on Safari.
-  ...(ownerState.margin === 'normal' && {
-    marginTop: 16,
-    marginBottom: 8,
-  }),
-  ...(ownerState.margin === 'dense' && {
-    marginTop: 8,
-    marginBottom: 4,
-  }),
-  ...(ownerState.fullWidth && {
-    width: '100%',
-  }),
-}));
+  variants: [
+    {
+      props: { margin: 'normal' },
+      style: {
+        marginTop: 16,
+        marginBottom: 8,
+      },
+    },
+    {
+      props: { margin: 'dense' },
+      style: {
+        marginTop: 8,
+        marginBottom: 4,
+      },
+    },
+    {
+      props: { fullWidth: true },
+      style: {
+        width: '100%',
+      },
+    },
+  ],
+});
 
 /**
  * Provides context such as filled/focused/error/required for form inputs.
